Install rst8 handler and prog to call and check it

+++ Makefile
@@ -63,19 +63,25 @@ rsx-c-prl: build
 	@echo "[RSX-C-PRL] Using z80asm to build dual ORG images (ASM+C)"
 	@if ! command -v zcc >/dev/null 2>&1; then echo "ERROR: zcc not in PATH"; exit 1; fi
 	@if ! command -v z80asm >/dev/null 2>&1; then echo "ERROR: z80asm not in PATH"; exit 1; fi
-	rm -f rsx_body_template.bin rsx_body_real.bin src/rsx_cfunc.o
+	rm -f rsx_body_template.bin rsx_body_real.bin src/rsx_cfunc.o src/pcw_rst8_c.o src/state.o
 
-	@echo "[RSX-C-PRL] Compiling C to object file"
+	@echo "Compiling C object files"
 	( cd src && zcc +cpm -c rsx_cfunc.c -o rsx_cfunc.o ) || exit 1
 	@if [ ! -f src/rsx_cfunc.o ]; then echo "ERROR: rsx_cfunc.o not produced"; exit 1; fi	
 
+	( cd src && zcc +cpm -compiler=sdcc -c pcw_rst8.c -o pcw_rst8_c.o ) || exit 1
+	@if [ ! -f src/pcw_rst8_c.o ]; then echo "ERROR: pcw_rst8_c.o not produced"; exit 1; fi	
+
+	( cd src && zcc +cpm  -c state.c -o state.o ) || exit 1
+	@if [ ! -f src/state.o ]; then echo "ERROR: state.o not produced"; exit 1; fi	
+
 	@echo "[RSX-C-PRL] Template (TEMPLATE macro -> ORG 0000h)"
-	( cd src && z80asm -b -DTEMPLATE rsx_body.asm rsx_cfunc.o ) || exit 1
+	( cd src && z80asm -b -DTEMPLATE rsx_body.asm  pcw_rst8.asm pcw_rst8_c.o state.o rsx_cfunc.o ) || exit 1
 	mv src/rsx_body.bin rsx_body_template.bin
 	@if [ ! -f rsx_body_template.bin ]; then echo "ERROR: rsx_body_template.bin not produced (template)"; exit 1; fi
 
 	@echo "[RSX-C-PRL] Real (default ORG 0100h)"
-	( cd src && z80asm -b rsx_body.asm rsx_cfunc.o ) || exit 1
+	( cd src && z80asm -b rsx_body.asm pcw_rst8.asm pcw_rst8_c.o state.o rsx_cfunc.o ) || exit 1
 	mv src/rsx_body.bin rsx_body_real.bin
 	@if [ ! -f rsx_body_real.bin ]; then echo "ERROR: rsx_body_real.bin not produced (real)"; exit 1; fi
 
@@ -117,7 +123,7 @@ tools/makeprl: tools/makeprl.c
 .PHONY: rsxchk
 rsxchk:
 	zcc +cpm -compiler=sdcc -clib=sdcc_ix -SO3 -vn -O2 \
-	  src/rsxchk.asm \
+	  src/rsxchk.c \
 	  -o RSXCHK -create-app
 
 build/gdbserver: $(GDBSERVER_C_OBJECTS) $(GDBSERVER_ASM_OBJECTS)
